Move tuple arguments related functions to Sample.hpp
